WIP: move files into core
