+++ packages/qwik/src/testing/element-fixture.ts
@@ -4,8 +4,8 @@ import { tryGetContext, type QContext } from '../core/state/context';
 import { normalizeOnProp } from '../core/state/listeners';
 import { getWrappingContainer, type PossibleEvents } from '../core/use/use-core';
 import { fromCamelToKebabCase } from '../core/util/case';
-import { getDomContainer } from '../serializer/client/dom-container';
-import { parseQRL } from '../serializer/shared-serialization';
+import { getDomContainer } from '../core/v2/client/dom-container';
+import { parseQRL } from '../core/v2/shared-serialization';
 import { createWindow } from './document';
 import { getTestPlatform } from './platform';
 import type { MockDocument, MockWindow } from './types';
